review:fix: headers are case-insensitive
