+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 * Fix #6906: Knative VolatileTime should be serialized as String
 * Fix #6930: Add support for Boolean enums in the java-generator
 * Fix #6886: Remove invalid JUnit 4 references
-* Fix #6917: Fabric 8 client does not authenticate correctly on openshift if the returned Location header is lower case
+* Fix #6917: Client does not authenticate correctly on OpenShift if the returned Location header is lower-case
 
 #### Improvements
 * Fix #6863: ensuring SerialExecutor does not throw RejectedExecutionException to prevent unnecessary error logs
