+++ httpclient-vertx/src/main/java/io/fabric8/kubernetes/client/vertx/VertxHttpRequest.java
@@ -16,7 +16,6 @@
 package io.fabric8.kubernetes.client.vertx;
 
 import io.fabric8.kubernetes.client.http.AsyncBody;
-import io.fabric8.kubernetes.client.http.HttpRequest;
 import io.fabric8.kubernetes.client.http.HttpResponse;
 import io.fabric8.kubernetes.client.http.StandardHttpRequest;
 import io.fabric8.kubernetes.client.http.StandardHttpRequest.BodyContent;
@@ -34,53 +33,10 @@ import java.nio.ByteBuffer;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Function;
 
 class VertxHttpRequest {
-
-  private static final class VertxHttpResponse implements HttpResponse<AsyncBody> {
-    private final AsyncBody result;
-    private final HttpClientResponse resp;
-    private final HttpRequest request;
-
-    private VertxHttpResponse(AsyncBody result, HttpClientResponse resp, HttpRequest request) {
-      this.result = result;
-      this.resp = resp;
-      this.request = request;
-    }
-
-    @Override
-    public int code() {
-      return resp.statusCode();
-    }
-
-    @Override
-    public AsyncBody body() {
-      return result;
-    }
-
-    @Override
-    public HttpRequest request() {
-      return request;
-    }
-
-    @Override
-    public Optional<HttpResponse<?>> previousResponse() {
-      return Optional.empty();
-    }
-
-    @Override
-    public List<String> headers(String key) {
-      return resp.headers().getAll(key);
-    }
-
-    @Override
-    public Map<String, List<String>> headers() {
-      return toHeadersMap(resp.headers());
-    }
-  }
 
   final Vertx vertx;
   private final RequestOptions options;
